requirements.txt
Flask==3.0.3
matplotlib==3.8.4
numba==0.59.1
numpy==1.24.3
pandas==2.2.2
Pillow==10.4.0
scikit_learn==1.4.2
tensorflow==2.13.1
tensorflow_intel==2.13.1
